move special into infer_type
